trebuie sa nu mai fac probleme seara ca ratez chestii mici si spamez EC

+++ 1D2DArrays/leet1277.cpp
@@ -16,13 +16,13 @@ public:
 
         prefix.resize(m + 1, vector<int>(n + 1, 0));
 
-        for(size_t i = 1; i <= (size_t)m; i++)
-            for(size_t j = 1; j <= (size_t)n; j++)
+        for(int i = 1; i <= m; i++)
+            for(int j = 1; j <= n; j++)
                 prefix[i][j] = matrix[i-1][j-1] + prefix[i-1][j] + prefix[i][j-1] - prefix[i-1][j-1];
         
         for(short int k = 1; k <= (short int)min(m, n); k++) {
-            for(size_t i = 1; i <= (size_t)(m-k+1); i++)
-                for(size_t j = 1; j <= (size_t)(n-k+1); j++)
+            for(int i = 1; i <= (m-k+1); i++)
+                for(int j = 1; j <= (n-k+1); j++)
                     if((prefix[i+k-1][j+k-1] 
                     - prefix[i-1][j+k-1] 
                     - prefix[i+k-1][j-1] 
@@ -37,8 +37,8 @@ int main() {
     int m, n;
     cin >> m >> n;
     vector<vector<int>> matrix(m, vector<int>(n));
-    for(size_t i = 0; i < m; i++) {
-        for (size_t j = 0; j < n; j++) {
+    for(int i = 0; i < m; i++) {
+        for (int j = 0; j < n; j++) {
             int val;
             cin >> val;
             matrix[i][j] = val;
